+++ yundun-bastionhost-20191209/alibabacloud_yundun_bastionhost20191209/models/_verify_instance_adauth_server_request.py
@@ -30,7 +30,6 @@ class VerifyInstanceADAuthServerRequest(DaraModel):
         self.instance_id = instance_id
         # This parameter is required.
         self.is_ssl = is_ssl
-        # This parameter is required.
         self.password = password
         # This parameter is required.
         self.port = port